requirements.txt
Flask==2.3.1
imessage-reader==0.5.0
Pillow==9.5.0
